Update trainer runs to use raw value + calculated, rather than a tuple
The tuple didn't quite work: we were using just the high 5 bits to 'approximate' the value. This worked correctly in all standard cases, but had issue with rounding. For example, `250` should be interpreted as 30 (and was correctly interpreted in the trainer team text box) but was instead being interpretted as 31 when looking at only the 5 high bits (as a tuple).

Replace this with a raw value shown in the hex content, with a calculated value (to 2 decimal points) shown in the table tool.

Also fixed some issues with Shiny's int-to-double calculated field change

+++ src/HexManiac.Core/ViewModels/Tools/CalculatedElementViewModel.cs
@@ -10,7 +10,7 @@ namespace HavenSoft.HexManiac.Core.ViewModels.Tools {
       private int dataStart;
 
       public string Name => segment.Name;
-      public int CalculatedValue {
+      public double CalculatedValue {
          get {
             var table = segment.Model.GetNextRun(dataStart) as ITableRun;
             if (table == null) return 0;
